fix(fedimint-client): api announcements not working if any peer failed
`join_all`, with an unbounded `retry` inside it, means that if any
of the peers has issues, the responses for other peers never get
processed.

This worked - somewhat - before, because the db write was done from
within `join_all` as well, but it would still lead to the whole refresh
loop getting stuck forever, etc. and probably was not a good idea.

The right way is to just fail, and process whatever responses did
succeed. The outer retry task will try again eventually anyway.