chore: update nixos module in the flake
Historically the module in the flake was first, and slightly different,
then we've managed to get module into nixpkgs, so flake version was left
as it was and forgotten about.

Since

https://github.com/NixOS/nixpkgs/pull/397967

is taking a long time to get any review, and might be stuck
indefinitely, update the module we export in our flake to the
nixpkgs PRs content, so they are the same.