fix(recurringd): avoid "holes" in invoice indices getting sync stuck
The main fix was choosing and incrementing the invoice index in the
same database transaction as saving the generated invoice. But since we
unfortunately can't also generate the invoice in the same dbtx some more
work was required.

In the old code, aborting the function would result in an unused index,
the problem with that was that the client always checks for the next
index and if none is present assumes no further invoices have been
generated. This obviously breaks in the presence of skipped indices.

In the new model the failure mode is changed: we'll never skip an index,
but rather generate two invoices with the same index if not careful
(since the invoice is still generated using a separate dbtx). This would
result in trying to add a second operation with the same operation id,
causing a panic. To avoid this the code now checks if there was an
aborted call to the `create_bolt11_invoice` function and repairs the DB
if so, only then it generates the actually requested invoice.