feat: allow guardian password changes
Changing the guardian password means both changing and re-encrypting the
private config on disk. This PR adds an API endpoint that allows a
guardian to trigger this. To avoid complexity this implementation
requires a restart of fedimintd for the change to take effect, thus
fedimintd shuts down after the password change and relies on the user or
their service manager (e.g. `systemd`) to initiate the restart.

Great care has been taken to make this process a) atomic and b) work in
the absence of a password.secret file, which we probably want to phase
out eventually. The design should ensure that every possible crash
during password change is recoverable.