chore: leniently parse e-cash without padding
Working with base64 data can be challanging, hence the earlier switch to
base64_urlsafe, which replaces some of the special characters with ones
not causing trouble in URLs. But equal signs, used at the end of both,
can still mess with copying and pasting. E.g. on iPhones it was reported
that the operating system tries to interpret the e-cash as math and
calculate some result, which messes things up.

This PR introduces lenient parsing such that we could drop the equal
signs at the end of e-cash in the future. Activation of that has to be
delayed since we guarantee stability for 4 versions. I would propose
a similar approach to the last change where it becomes available as the
alt encoding first and later becomes the standard.