fix: improve `refresh_peers_api_versions`
If we already had a response and we are not planning to retry,
we can increment the "number of responses".

I don't think this had any negative user-visible consequences,
as client is not blocked on this counter anyway, but seems
technically incorrect.

(cherry picked from commit 0c6ce5d8a45d50708f18f803d7db9eef49faec9a)