chore(ci): on failure output logs from services with a prefix
When scrolling in the github UI to quickly evaluate failure
I have a hard time finding the beginning and the end and
figoure out which output I'm looking at. By prefixing them
with a short string I'm hoping it will be easier to tell
when scrolling what are we looking at.