refactor(client): improve ways client events can be consumed
This generalized tracking to use a `EventLogTracker` trait,
so it can be implemented more flexibly, e.g. allowing storing
position in prefixed db namespaces, etc. The old method
of using a key was not very practical, as keys itself
are quite static (can't generate prefix at runtime).

A prefix was allocated for built-in event log tracker
for the main downstream application driving the Client.

Some comments and renamed were done to make handling
event log easier.

I have some ideas for better naming system:

https://github.com/fedimint/fedimint/issues/7940

but that is going to be a larger rename change, so
maybe in a different PR.

To demonstrate and test how to use `Client::handle_events`
and new CLI command was added:

```
fm-cli dev test-event-log-handling
```