fix: larger delay before shutting down api
I think we've managed to still hit this flake once,
though it seems to be much, much more rare now.
Let's increase the sleep by an order of magnitude
still. I don't want to go to high because this is
slowing down every dkg we use in our test suite, and
seems like 10ms is already almost completely reliable.