chore(ci): skip building some thing after PR lands to master
In particular, building and publishg releases for every commit in master
branch is *very* heavy.

https://github.com/fedimint/fedimint/actions/runs/19973840069/job/57284938310

I think we want to do it only on release branches and tags,
where people are expected to might want to try the artefacts.

The other steps are more about dev, and probably caching
will make them relatively fast, but there is really no point in running
them - merge queue running serially, guarantees that everything
was verified already, so we just waste couple of minutes of
runner's time where it could be doing something more useful.