chore(test): more informative error message
In

https://github.com/fedimint/fedimint/actions/runs/19976014735/job/57292258559

I see a flake that errors out with:

```
Error: Not all announcements ready: 4
```

which is weird, because we do run with 4 peers, right? So
what else would it expect?

I suspect the condition there should be `<` than instead of `!=`,
but I can't find where in the code we would actually have a different
peer number, so a simple approach is to print that number
to verify.