feat(dev): introduce SelfCI
SelfCI is a little project I'm working on:

https://app.radicle.xyz/nodes/radicle.dpc.pw/rad%3Az2tDzYbAXxTQEKTGFVwiJPajkbeDU

If you wouldn't mind these little files here would
allow me to use it (locally) with Fedimint and
ensure that it works well for projects like
ours. For now if anyone would like to try it out,
they will have to set it up themselves, and since
it's still very immature I can't guarantee it will
be a smooth ride.

For now it's very handy e.g. for being able to locally
verify changes kind of like CI would do, but being
free to change the local workdir to a different branch
while it is running.